requirements.txt
Django==2.2.*
psycopg2-binary>=2.8
djangorestframework==3.12.*
djangorestframework-simplejwt==4.6.*
djoser==2.1.*
django-cors-headers==3.7.*
django-filter==2.4.*